reorganized to remove unnecessary maze mod
reorganized to remove unnecessary maze mod
